// src/providers/cmc/types/cmc-enum.type.ts
export enum CmcEnvironmenType {
  SANDBOX = 'sandbox',
  PRODUCTION = 'production',
}

export enum CmcCategory {
  CRYPTOCURRENCY = 'cryptocurrency',
  EXCHANGE = 'exchange',
  GLOBAL_METRICS = 'global-metrics',
  TOOLS = 'tools',
  BLOCKCHAIN = 'blockchain',
  FIAT = 'fiat',
  PARTNERS = 'partners',
  KEY = 'key',
  CONTENT = 'content',
}